networking: cleanup/refactor common code
Adds a setting for enabling a development
network interface. This is useful for testing
the stack in an environment where the physical
cisco/pfsense environment is not available.

The idea is that eth0 will be attached to a
network with no gateway that all of the vms
of the stack will be on in the production
subnet 172.16.2.0/24. We then use dhcp and SLAAC
on the eth1 interface simply to get a default
route to the internet and also to ssh into
the machines for deployment process.