samba: fix breakage causing RPC unavaiable errors
I'm not sure why this was committed in the first place
but it seems to have caused major breakage.

With these options, ADCS and GPO Management both
indicate that the RPC Server is down or something like that

Samba logs indicated an NT_STATUS_INVALID_SID error.

These 2 posts I found online led me to finding this as the culprit:
https://samba.samba.narkive.com/mGVMblr9/nt-status-invalid-sid-in-a-sdc
https://bugzilla.samba.org/show_bug.cgi?id=12410